Refactor the private storage class to be more generic for projects
