+++ opentech/apply/projects/models.py
@@ -12,6 +12,9 @@ from django.db import models
 from django.urls import reverse
 from django.utils.translation import ugettext as _
 
+from opentech.apply.utils.storage import private_storage
+
+
 logger = logging.getLogger(__name__)
 
 
@@ -37,7 +40,7 @@ class PacketFile(models.Model):
     project = models.ForeignKey("Project", on_delete=models.CASCADE, related_name="packet_files")
 
     title = models.TextField()
-    document = models.FileField(upload_to=document_path)
+    document = models.FileField(upload_to=document_path, storage=private_storage())
 
     def __str__(self):
         return f'Project file: {self.title}'
