+++ opentech/settings/dev.py
@@ -3,6 +3,8 @@ from .base import *  # noqa
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
+PROJECTS_ENABLED = True
+
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = 'CHANGEME!!!'
