+++ opentech/storage_backends.py
@@ -1,7 +1,6 @@
 from urllib import parse
 
 from django.conf import settings
-from django.urls import reverse
 from django.utils.encoding import filepath_to_uri
 from storages.backends.s3boto3 import S3Boto3Storage
 
@@ -29,21 +28,8 @@ class PrivateMediaStorage(S3Boto3Storage):
     file_overwrite = False
     querystring_auth = True
     url_protocol = 'https:'
-    is_submission = False
 
     def url(self, name, parameters=None, expire=None):
-        if self.is_submission:
-            try:
-                name_parts = name.split('/')
-                return reverse(
-                    'apply:submissions:serve_private_media', kwargs={
-                        'submission_id': name_parts[1], 'field_id': name_parts[2],
-                        'file_name': name_parts[3]
-                    }
-                )
-            except IndexError:
-                pass
-
         url = super().url(name, parameters, expire)
 
         if hasattr(settings, 'AWS_PRIVATE_CUSTOM_DOMAIN'):
